+++ Framework/Api/ViewEngine.cs
@@ -92,6 +92,19 @@ public class ViewEngine : IViewEngine
     }
 
     /// <inheritdoc />
+    public void PreloadAssets()
+    {
+        Task.Run(assetRegistry.PreloadAsync)
+            .ContinueWith(t =>
+            {
+                if (t.IsFaulted)
+                {
+                    Logger.Log($"An unexpected error occurred during asset preloading: {t.Exception}", LogLevel.Error);
+                }
+            });
+    }
+
+    /// <inheritdoc />
     public void RegisterCustomData(string assetPrefix, string modDirectory)
     {
         assetRegistry.RegisterCustomData(assetPrefix, modDirectory);